Safeguard naming character
Prohibits user from creating a character with the name of an existing character. Also does not allow a user to continue if there is no name in the name field.

Added a feature to make the first letter of the character's name to be automatically converted to uppercase when proceeding to the next step.